requirements.txt
bluepy
pycrypto
curses-menu
uvicorn[standard]
fastapi
paho-mqtt
